Misc. bug fixes and some new widgets
