+++ pyegeria/commands/cli/egeria.py
@@ -29,7 +29,6 @@ from pyegeria.commands.cat.list_cert_types import display_certifications
 from pyegeria.commands.cat.list_terms import display_glossary_terms
 from pyegeria.commands.cat.list_projects import display_project_list
 from pyegeria.commands.tech.list_anchored_elements import display_anchored_elements
-from pyegeria.commands.tech.list_elements_x import list_elements_x
 from pyegeria.commands.tech.list_relationships import list_relationships
 from pyegeria.commands.cat.list_tech_types import display_tech_types
 from pyegeria.commands.cat.list_todos import display_to_dos as list_todos
@@ -99,12 +98,19 @@ from pyegeria.commands.tech.list_asset_types import display_asset_types
 from pyegeria.commands.tech.list_elements_for_classification import (
     list_classified_elements,
 )
-from pyegeria.commands.tech.list_elements import list_elements
+from pyegeria.commands.tech.list_all_om_type_elements import list_elements
+from pyegeria.commands.tech.list_all_om_type_elements_x import list_elements_x
+from pyegeria.commands.tech.list_elements_by_classification_by_property_value import find_elements_by_classification_by_prop_value
+from pyegeria.commands.tech.list_elements_by_property_value import find_elements_by_prop_value
+from pyegeria.commands.tech.list_elements_by_property_value_x import find_elements_by_prop_value_x
+from pyegeria.commands.tech.list_related_elements_with_prop_value import list_related_elements_with_prop_value
+
+
 from pyegeria.commands.tech.list_registered_services import display_registered_svcs
 from pyegeria.commands.tech.list_related_specification import (
     display_related_specification,
 )
-from pyegeria.commands.tech.list_related_elements import list_related_elements
+from pyegeria.commands.tech.list_all_related_elements import list_related_elements
 
 from pyegeria.commands.tech.list_relationship_types import display_relationship_types
 from pyegeria.commands.tech.list_tech_templates import display_templates_spec
@@ -382,7 +388,7 @@ def show_guid_info(ctx, guid):
 @show_elements.command("anchored_elements")
 @click.pass_context
 @click.option(
-    "--search-string",
+    "--property_value",
     default="DeployedDatabaseSchema",
     help="value we are searching for",
 )
@@ -391,8 +397,8 @@ def show_guid_info(ctx, guid):
     default="anchorTypeName",
     help="List of properties we are searching",
 )
-def list_anchored_elements(ctx, search_string: str, prop_list: str):
-    """List elements with the specified properties"""
+def list_anchored_elements(ctx, property_value: str, prop_list: str):
+    """List anchored elements with the specified properties"""
     c = ctx.obj
     if type(prop_list) is str:
         property_names = prop_list.split(",")
@@ -403,13 +409,198 @@ def list_anchored_elements(ctx, search_string: str, prop_list: str):
         print(f"\nError --> Invalid property list - must be a string or list")
         sys.exit(4)
     display_anchored_elements(
-        search_string,
+        property_value,
         [prop_list],
         c.view_server,
         c.view_server_url,
         c.userid,
         c.password,
         c.timeout,
+        c.jupyter,
+        c.width,
+    )
+
+@show_elements.command("elements-by-classification")
+@click.option(
+    "--om-type",
+    default="Referenceable",
+    help="Open Metadata type to filter by",
+)
+@click.option(
+    "--classification",
+    default="GovernanceProject",
+    help="Classification to filter byt",
+)
+@click.pass_context
+def show_elements_by_classification(ctx, om_type, classification):
+    """Show elements by classification"""
+    c = ctx.obj
+    list_classified_elements(
+        om_type,
+        classification,
+        c.view_server,
+        c.view_server_url,
+        c.userid,
+        c.password,
+        c.jupyter,
+        c.width,
+    )
+
+@show_elements.command("elements-by-classification-by-prop-value")
+@click.option(
+    "--classification",
+    default="GovernanceProject",
+    help="Classification to filter by",
+    )
+@click.option(
+    "--property_value",
+    help="Property value to filter by",
+)
+@click.option(
+    "--property_names",
+    help="List of properties to search by",
+)
+@click.option(
+    "--om-type",
+    default="Referenceable",
+    help="Open Metadata type to filter by",
+)
+@click.pass_context
+def show_elements_by_classification_by_prop(ctx,  classification, property_value, property_names, om_type):
+    """Show elements by classification and property value"""
+    c = ctx.obj
+    find_elements_by_classification_by_prop_value(
+        om_type,
+        classification,
+        property_value,
+        [property_names],
+        c.view_server,
+        c.view_server_url,
+        c.userid,
+        c.password,
+        c.jupyter,
+        c.width,
+    )
+
+@show_elements.command("elements-by-prop-value")
+@click.option(
+    "--property_value",
+    help="Property value to filter by",
+)
+@click.option(
+    "--property_names",
+    help="List of properties to search by",
+)
+@click.option(
+    "--om-type",
+    default="Referenceable",
+    help="Open Metadata type to filter by",
+)
+@click.option(
+    "--extended",
+    is_flag=True,
+    default=False,
+    help="If True, feedback information is displayed",
+)
+@click.pass_context
+def show_elements_by_classification_by_prop(ctx,property_value, property_names, om_type, extended):
+    """Show elements by classification and property value"""
+    c = ctx.obj
+    if extended:
+        find_elements_by_prop_value_x(
+            om_type,
+            property_value,
+            [property_names],
+            c.view_server,
+            c.view_server_url,
+            c.userid,
+            c.password,
+            c.jupyter,
+            c.width,
+        )
+    else:
+        find_elements_by_prop_value(
+            om_type,
+            property_value,
+            [property_names],
+            c.view_server,
+            c.view_server_url,
+            c.userid,
+            c.password,
+            c.jupyter,
+            c.width,
+            )
+
+@show_elements.command("related-elements")
+@click.option(
+    "--element-guid",
+    help="GUID of the Element to navigate from.",
+)
+@click.option(
+    "--om-type",
+    default="Referenceable",
+    help="Open metadata type to filter by.",
+)
+@click.option(
+    "--rel-type",
+    default="Certification",
+    help="Relationship type to follow.",
+)
+@click.pass_context
+def show_related_elements(ctx, element_guid, om_type, rel_type):
+    """Show all elements related to specified guid"""
+    c = ctx.obj
+    list_related_elements(
+        element_guid,
+        om_type,
+        rel_type,
+        c.view_server,
+        c.view_server_url,
+        c.userid,
+        c.password,
+        c.jupyter,
+        c.width,
+    )
+
+
+@show_elements.command("related-elements_by_prop")
+@click.option(
+    "--element-guid",
+    help="GUID of the Element to navigate from.",
+)
+@click.option(
+    "--rel-type",
+    default="Certification",
+    help="Relationship type to follow.",
+)
+@click.option(
+    "--property_value",
+    help="Property value to filter by",
+)
+@click.option(
+    "--property_names",
+    help="List of properties to search by",
+)
+@click.option(
+    "--om-type",
+    default="Referenceable",
+    help="Open metadata type to filter by.",
+)
+
+@click.pass_context
+def show_related_elements(ctx, element_guid, rel_type, property_value, property_names, om_type):
+    """Show elements related to specified guid and property value"""
+    c = ctx.obj
+    list_related_elements_with_prop_value(
+        element_guid,
+        rel_type,
+        property_value,
+        [property_names],
+        om_type,
+        c.view_server,
+        c.view_server_url,
+        c.userid,
+        c.password,
         c.jupyter,
         c.width,
     )
@@ -534,7 +725,7 @@ def show_registered_services(ctx, services):
 @show_tech_info.command("relationship-types")
 @click.option(
     "--om-type",
-    default="AssetOwner",
+    default="Referenceable",
     help="Relationship type to get information about",
 )
 @click.pass_context
@@ -553,63 +744,7 @@ def show_relationship_types(ctx, om_type):
     )
 
 
-@show_elements.command("elements-by-classification")
-@click.option(
-    "--om-type",
-    default="Project",
-    help="Open Metadata type to filter by",
-)
-@click.option(
-    "--classification",
-    default="GovernanceProject",
-    help="Classification to filter byt",
-)
-@click.pass_context
-def show_elements_by_classification(ctx, om_type, classification):
-    """Show elements by classification"""
-    c = ctx.obj
-    list_classified_elements(
-        om_type,
-        classification,
-        c.view_server,
-        c.view_server_url,
-        c.userid,
-        c.password,
-        c.jupyter,
-        c.width,
-    )
 
-
-@show_elements.command("related-elements")
-@click.option(
-    "--element-guid",
-    help="GUID of the Element to navigate from.",
-)
-@click.option(
-    "--om-type",
-    default="Project",
-    help="Open metadata type to filter by.",
-)
-@click.option(
-    "--rel-type",
-    default="Certification",
-    help="Relationship type to follow.",
-)
-@click.pass_context
-def show_related_elements(ctx, element_guid, om_type, rel_type):
-    """Show elements related to specified guid"""
-    c = ctx.obj
-    list_related_elements(
-        element_guid,
-        om_type,
-        rel_type,
-        c.view_server,
-        c.view_server_url,
-        c.userid,
-        c.password,
-        c.jupyter,
-        c.width,
-    )
 
 
 @show_tech_type.command("template-spec")
@@ -698,9 +833,9 @@ def valid_metadata_values(ctx, property, type_name):
     default=False,
     help="If True, feedback information is displayed",
 )
-@click.option("--om_type", default="Organization", help="Metadata type to query")
-def list_element_info(ctx, om_type, extended):
-    """Display elements of a specific Open Metadata Type"""
+@click.option("--om_type", default="Referenceable", help="Metadata type to query")
+def list_all_om_type_elements(ctx, om_type, extended):
+    """Display all elements of a specific Open Metadata Type"""
     c = ctx.obj
     if extended:
         list_elements_x(
@@ -710,7 +845,7 @@ def list_element_info(ctx, om_type, extended):
             c.userid,
             c.password,
             c.jupyter,
-            c.width,
+            c.width
         )
     else:
         list_elements(
@@ -720,7 +855,7 @@ def list_element_info(ctx, om_type, extended):
             c.userid,
             c.password,
             c.jupyter,
-            c.width,
+            c.width
         )
 
 
@@ -742,7 +877,7 @@ def list_element_info(ctx):
 
 @show_elements.command("get-elements")
 @click.pass_context
-@click.option("--om_type", default="Project", help="Metadata type to query")
+@click.option("--om_type", default="Referenceable", help="Metadata type to query")
 def get_element_info(ctx, om_type):
     """Display a table of elements of an Open Metadata Type"""
     c = ctx.obj
@@ -1620,18 +1755,14 @@ def engine_host(ctx):
 @click.option(
     "--list", is_flag=True, default=False, help="If True, a paged list will be shown"
 )
-@click.option(
-    "--engine-host",
-    default="engine-host",
-    help="Name of the Engine Host to get status for",
-)
+
 @click.pass_context
 def gov_eng_status(ctx, engine_list, engine_host, list):
     """Display engine-host status information"""
     c = ctx.obj
     display_gov_eng_status(
         [engine_list],
-        engine_host=engine_host,
+        engine_host=c.engine_host,
         view_server=c.view_server,
         url=c.view_server_url,
         username=c.userid,
