+++ pyegeria/commands/cli/egeria_ops.py
@@ -292,11 +292,7 @@ def engine_host(ctx):
     default="*",
     help="Enter the list of connectors you are interested in or ['*'] for all",
 )
-@click.option(
-    "--engine-host",
-    default="engine-host",
-    help="Name of the Engine Host to get status for",
-)
+
 @click.option(
     "--list", is_flag=True, default=False, help="If True, a paged list will be shown"
 )
@@ -306,7 +302,7 @@ def gov_eng_status(ctx, engine_list, engine_host, list):
     c = ctx.obj
     display_gov_eng_status(
         [engine_list],
-        engine_host=engine_host,
+        engine_host=c.engine_host,
         view_server=c.view_server,
         url=c.view_server_url,
         username=c.userid,
