+++ pyegeria/commands/tech/list_elements_by_property_value.py
@@ -14,8 +14,6 @@ from pyegeria import (
     PropertyServerException,
     UserNotAuthorizedException,
     print_exception_response,
-    # ClassificationManager,
-    # FeedbackManager,
     EgeriaTech,
 )
 
@@ -37,14 +35,16 @@ EGERIA_JUPYTER = bool(os.environ.get("EGERIA_JUPYTER", "False"))
 EGERIA_WIDTH = int(os.environ.get("EGERIA_WIDTH", "200"))
 
 
-def list_elements_x(
-    om_type: str,
-    server: str,
-    url: str,
-    username: str,
-    password: str,
-    jupyter: bool = EGERIA_JUPYTER,
-    width: int = EGERIA_WIDTH,
+def find_elements_by_prop_value(
+        om_type: str,
+        property_value: str,
+        property_names: [str],
+        server: str,
+        url: str,
+        username: str,
+        password: str,
+        jupyter: bool = EGERIA_JUPYTER,
+        width: int = EGERIA_WIDTH,
 ):
     c_client = EgeriaTech(server, url, user_id=username, user_pwd=password)
     token = c_client.create_egeria_bearer_token()
@@ -55,12 +55,14 @@ def list_elements_x(
             f"The type name '{om_type}' is not known to the Egeria platform at {url} - {server}"
         )
         sys.exit(1)
-    elements = c_client.get_elements(om_type)
+    elements = c_client.find_elements_by_property_value(property_value, property_names,
+                                                        om_type
+                                                        )
 
     def generate_table() -> Table:
         """Make a new table."""
         table = Table(
-            caption=f"Metadata Elements for: {url} - {server} @ {time.asctime()}",
+            caption=f"Find Metadata Elements for: {url} - {server} @ {time.asctime()}",
             style="bold bright_white on black",
             row_styles=["bold bright_white on black"],
             header_style="white on dark_blue",
@@ -68,7 +70,8 @@ def list_elements_x(
             caption_style="white on black",
             show_lines=True,
             box=box.ROUNDED,
-            title=f"Elements for Open Metadata Type: '{om_type}' ",
+            title=f"Elements for Open Metadata Type: {om_type}, property value: {property_value}, "
+                  f"properties: {property_names}",
             expand=True,
             width=width,
         )
@@ -79,8 +82,7 @@ def list_elements_x(
         table.add_column("Home Store")
         table.add_column("GUID", width=38, no_wrap=True)
         table.add_column("Properties")
-        table.add_column("Feedback", min_width=30)
-        table.add_column("Public Comments")
+        table.add_column("Classifications")
 
         if type(elements) is list:
             for element in elements:
@@ -88,48 +90,29 @@ def list_elements_x(
                 el_q_name = element["properties"].get("qualifiedName", "---")
                 el_type = header["type"]["typeName"]
                 el_home = header["origin"]["homeMetadataCollectionName"]
-                el_create_time = header["versions"]["createTime"][:-18]
+                el_create_time = header["versions"]["createTime"][:-10]
                 el_guid = header["guid"]
+                el_class = header.get("classifications", "---")
 
                 el_props_md = ""
                 for prop in element["properties"].keys():
                     el_props_md += f"* **{prop}**: {element['properties'][prop]}\n"
-
                 el_props_out = Markdown(el_props_md)
 
-                tags = c_client.get_attached_tags(el_guid)
-                tags_md = "Tags:\n"
-                if type(tags) is list:
-                    for tag in tags:
-                        tags_md += (
-                            f"* tag: {tag.get('name','')}\n"
-                            f"\t description: {tag.get('description','---')}\n"
-                            f"\t assigned by: {tag.get('user','---')}\n"
+                c_md = ""
+                if type(el_class) is list:
+                    for classification in el_class:
+                        classification_name = classification.get(
+                            "classificationName", "---"
                         )
-
-                else:
-                    tags_md = ""
-
-                likes = c_client.get_attached_likes(el_guid)
-                likes_md = "Likes:\b"
-                if type(likes) is list:
-                    for like in likes:
-                        likes_md += (
-                            f"* tag: {like['name']}\n"
-                            f"* description: {like['description']}\n"
-                            f"* assigned by: {like['user']}\n"
-                            f"\n"
+                        c_md = f"* **{classification_name}**\n"
+                        class_props = classification.get(
+                            "classificationProperties", "---"
                         )
-
-                else:
-                    likes_md = ""
-
-                if len(tags_md) > 0 and len(likes_md) > 0:
-                    feedback_out = f"{tags_md}\n --- \n{likes_md}"
-                else:
-                    feedback_out = ""
-
-                comments_out = " "
+                        if type(class_props) is dict:
+                            for prop in class_props.keys():
+                                c_md += f"  * **{prop}**: {class_props[prop]}\n"
+                c_md_out = Markdown(c_md)
 
                 table.add_row(
                     el_q_name,
@@ -138,8 +121,7 @@ def list_elements_x(
                     el_home,
                     el_guid,
                     el_props_out,
-                    feedback_out,
-                    comments_out,
+                    c_md_out,
                 )
 
             return table
@@ -180,9 +162,11 @@ def main():
 
     try:
         om_type = Prompt.ask(
-            "Enter the Open Metadata Type to find elements of:", default="GlossaryTerm"
+            "Enter the Open Metadata Type to find elements of", default="Referenceable"
         )
-        list_elements_x(om_type, server, url, userid, password)
+        property_value = Prompt.ask("Enter the property value to search for")
+        property_names = Prompt.ask("Enter a comma seperated list of properties to search")
+        find_elements_by_prop_value(om_type, property_value, [property_names], server, url, userid, password)
     except KeyboardInterrupt:
         pass
